Move convert to bitmap into bitmap helper
